+++ packages/core/src/Matcher.ts
@@ -2,6 +2,7 @@ import React from 'react';
 import { MatchCallback, MatchResponse, Node, ChildrenNode } from './types';
 
 export interface MatcherInterface<T> {
+  greedy?: boolean;
   inverseName: string;
   propName: string;
   asTag(): string;
@@ -13,6 +14,8 @@ export interface MatcherInterface<T> {
 
 export default abstract class Matcher<Props extends object = {}, Options extends object = {}>
   implements MatcherInterface<Props> {
+  greedy: boolean = false;
+
   options: Options;
 
   propName: string;
@@ -65,6 +68,7 @@ export default abstract class Matcher<Props extends object = {}, Options extends
     string: string,
     pattern: string | RegExp,
     callback: MatchCallback<T>,
+    isVoid: boolean = false,
   ): MatchResponse<T> | null {
     const match = string.match(pattern instanceof RegExp ? pattern : new RegExp(pattern, 'i'));
 
@@ -74,6 +78,7 @@ export default abstract class Matcher<Props extends object = {}, Options extends
 
     return {
       match: match[0],
+      void: isVoid,
       ...callback(match),
       index: match.index!,
       length: match[0].length,